updated code

+++ TorchClassifier/Datasetutil/Visutil.py
@@ -27,7 +27,7 @@ def visbatchimage(images, labels, classes):
     fig = plt.figure(figsize=(25, 4))
     # display 20 images
     for idx in np.arange(20):
-        ax = fig.add_subplot(2, 20/2, idx+1, xticks=[], yticks=[])
+        ax = fig.add_subplot(2, int(20/2), idx+1, xticks=[], yticks=[])
         imshowhelper(images[idx])
         ax.set_title(classes[labels[idx]])
     fig.savefig('./outputs/torchonebatchimage.png')
@@ -38,7 +38,7 @@ def vistestresult(images, labels, preds, classes, path='./outputs/'):
     # plot the images in the batch, along with predicted and true labels
     fig = plt.figure(figsize=(25, 4))
     for idx in np.arange(20):
-        ax = fig.add_subplot(2, 20/2, idx+1, xticks=[], yticks=[])
+        ax = fig.add_subplot(2, int(20/2), idx+1, xticks=[], yticks=[])
         imshowhelper(images.cpu()[idx])
         ax.set_title("{} ({})".format(classes[preds[idx]], classes[labels[idx]]), color=("green" if preds[idx]==labels[idx].item() else "red"))
     figsavepath=os.path.join(path, 'torchtestresultimage.png')
